+++ package.json
@@ -26,18 +26,6 @@
     "@trpc/next": "^9.27.0",
     "@trpc/react": "^9.27.0",
     "@trpc/server": "^9.27.0",
-    "@udecode/plate-autoformat": "^16.5.0",
-    "@udecode/plate-basic-marks": "^16.5.0",
-    "@udecode/plate-block-quote": "^16.5.0",
-    "@udecode/plate-break": "^16.5.0",
-    "@udecode/plate-code-block": "^16.5.0",
-    "@udecode/plate-common": "^7.0.2",
-    "@udecode/plate-core": "^16.5.0",
-    "@udecode/plate-heading": "^16.5.0",
-    "@udecode/plate-list": "^16.6.1",
-    "@udecode/plate-paragraph": "^16.5.0",
-    "@udecode/plate-reset-node": "^16.5.0",
-    "@udecode/plate-trailing-block": "^16.5.0",
     "daisyui": "^2.22.0",
     "dayjs": "^1.11.5",
     "jotai": "^1.8.4",
@@ -45,15 +33,9 @@
     "next-auth": "^4.10.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.34.2",
     "react-hot-toast": "^2.4.0",
     "react-icons": "^4.4.0",
     "react-query": "3.39.2",
-    "slate": "^0.82.1",
-    "slate-history": "^0.66.0",
-    "slate-hyperscript": "^0.77.0",
-    "slate-react": "^0.82.1",
-    "styled-components": "^5.3.5",
     "superjson": "^1.9.1",
     "zod": "^3.17.3"
   },
