+++ src/components/editor/IconsToolbar.tsx
@@ -1,14 +1,6 @@
-import {
-  MARK_BOLD,
-  MARK_CODE,
-  MARK_ITALIC,
-  MARK_STRIKETHROUGH,
-  MARK_UNDERLINE,
-} from '@udecode/plate-basic-marks'
-import { ELEMENT_BLOCKQUOTE } from '@udecode/plate-block-quote'
-import { ELEMENT_CODE_BLOCK } from '@udecode/plate-code-block'
-import { ELEMENT_OL, ELEMENT_UL } from '@udecode/plate-list'
-import type { ReactNode } from 'react'
+import { clsx } from 'clsx'
+import { useAtomValue } from 'jotai'
+import { ReactNode } from 'react'
 import { BiCodeBlock } from 'react-icons/bi'
 import {
   MdCode,
@@ -19,60 +11,173 @@ import {
   MdFormatQuote,
   MdFormatStrikethrough,
   MdFormatUnderlined,
-  MdMenu,
+  MdRedo,
+  MdUndo,
 } from 'react-icons/md'
-import BlockIconButton from './BlockIconButton'
-import MarkIconButton from './MarkIconButton'
+import { twMerge } from 'tailwind-merge'
+import MenuButton from '../MenuButton'
+import { editorAtom, toolbarForcedAtom } from './TwinEditor'
 
-const IconsToolbar = ({ children }: { children?: ReactNode }) => (
-  <nav className='sticky top-0 z-10 mb-4 bg-base-100 px-4 pt-4 lg:px-8'>
-    <div className='mb-6 flex items-center gap-4'>
-      <label tabIndex={0} htmlFor='sidebar' className='btn-icon lg:hidden'>
-        <MdMenu className='h-6 w-6' />
-      </label>
-      <div className='flex flex-1 flex-nowrap overflow-y-auto'>
-        <div className='btn-group flex-nowrap rounded-lg bg-base-content/10'>
-          <BlockIconButton type={ELEMENT_UL} title='Bulleted List'>
-            <MdFormatListBulleted className='h-6 w-6' />
-          </BlockIconButton>
-          <BlockIconButton type={ELEMENT_OL} title='Numbered List'>
-            <MdFormatListNumbered className='h-6 w-6' />
-          </BlockIconButton>
-          <BlockIconButton
-            type={ELEMENT_BLOCKQUOTE}
-            title='Blockquote (ctrl+shift+.)'
-          >
-            <MdFormatQuote className='h-6 w-6' />
-          </BlockIconButton>
-          <BlockIconButton type={ELEMENT_CODE_BLOCK} title='Code Block'>
-            <BiCodeBlock className='h-6 w-6' />
-          </BlockIconButton>
-          <div className='w-[0.125rem] bg-base-content/30' />
-          <MarkIconButton type={MARK_BOLD} title='Bold (ctrl+b)'>
-            <MdFormatBold className='h-6 w-6' />
-          </MarkIconButton>
-          <MarkIconButton type={MARK_ITALIC} title='Italic (ctrl+i)'>
-            <MdFormatItalic className='h-6 w-6' />
-          </MarkIconButton>
-          <MarkIconButton type={MARK_UNDERLINE} title='Underline (ctrl+u)'>
-            <MdFormatUnderlined className='h-6 w-6' />
-          </MarkIconButton>
-          <MarkIconButton
-            type={MARK_STRIKETHROUGH}
-            title='Strikethrough (ctrl+shift+x)'
-          >
-            <MdFormatStrikethrough className='h-6 w-6' />
-          </MarkIconButton>
-          <MarkIconButton type={MARK_CODE} title='Inline code (ctrl+e)'>
-            <MdCode className='h-6 w-6' />
-          </MarkIconButton>
+const IconsToolbar = ({ children }: { children?: ReactNode }) => {
+  useAtomValue(toolbarForcedAtom)
+  const editor = useAtomValue(editorAtom)
+
+  return (
+    <nav className='sticky top-0 z-10 mb-4 bg-base-100 px-4 pt-4 lg:px-8'>
+      <div className='mb-6 flex items-center gap-4'>
+        <MenuButton />
+
+        <div className='flex flex-1 flex-nowrap overflow-y-auto'>
+          <div className='btn-group flex-nowrap rounded-lg bg-base-content/10'>
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('bulletList'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleBulletList().run()}
+              title='Bulleted List'
+            >
+              <MdFormatListBulleted className='h-6 w-6' />
+            </button>
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('orderedList'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleOrderedList().run()}
+              title='Numbered List'
+            >
+              <MdFormatListNumbered className='h-6 w-6' />
+            </button>
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('blockquote'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleBlockquote().run()}
+              title='Blockquote (ctrl+shift+.)'
+            >
+              <MdFormatQuote className='h-6 w-6' />
+            </button>
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('codeBlock'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleCodeBlock().run()}
+              title='Code Block'
+            >
+              <BiCodeBlock className='h-6 w-6' />
+            </button>
+
+            {/* Vertical divider */}
+            <div className='w-[0.125rem] bg-base-content/30' />
+
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('bold'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleBold().run()}
+              title='Bold (ctrl+b)'
+            >
+              <MdFormatBold className='h-6 w-6' />
+            </button>
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('italic'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleItalic().run()}
+              title='Italic (ctrl+i)'
+            >
+              <MdFormatItalic className='h-6 w-6' />
+            </button>
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('underline'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleUnderline().run()}
+              title='Underline (ctrl+u)'
+            >
+              <MdFormatUnderlined className='h-6 w-6' />
+            </button>
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('strike'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleStrike().run()}
+              title='Strikethrough (ctrl+shift+x)'
+            >
+              <MdFormatStrikethrough className='h-6 w-6' />
+            </button>
+            <button
+              disabled={!editor}
+              className={twMerge(
+                'btn-icon',
+                clsx({
+                  'btn-active': editor?.isActive('code'),
+                })
+              )}
+              onClick={() => editor?.chain().focus().toggleCode().run()}
+              title='Inline code (ctrl+e)'
+            >
+              <MdCode className='h-6 w-6' />
+            </button>
+
+            {/* Vertical divider */}
+            <div className='w-[0.125rem] bg-base-content/30' />
+
+            <button
+              disabled={!editor || !editor?.can().chain().focus().undo().run()}
+              className='btn-icon'
+              onClick={() => editor?.chain().focus().undo().run()}
+              title='Undo (ctrl+z)'
+            >
+              <MdUndo className='h-6 w-6' />
+            </button>
+            <button
+              disabled={!editor || !editor?.can().chain().focus().redo().run()}
+              className='btn-icon'
+              onClick={() => editor?.chain().focus().redo().run()}
+              title='Redo (ctrl+shift+z)'
+            >
+              <MdRedo className='h-6 w-6' />
+            </button>
+          </div>
         </div>
-      </div>
 
-      {children}
-    </div>
-    <div className='divider m-0 h-fit' />
-  </nav>
-)
+        {children}
+      </div>
+      <div className='divider m-0 h-fit' />
+    </nav>
+  )
+}
 
 export default IconsToolbar
