+++ src/components/editor/SmallEditor.tsx
