+++ src/pages/edit/[id].tsx
@@ -6,10 +6,10 @@ import { useMemo } from 'react'
 import { toast } from 'react-hot-toast'
 import { MdAdd } from 'react-icons/md'
 import { twMerge } from 'tailwind-merge'
+import IconsToolbar from '../../components/editor/IconsToolbar'
+import TwinEditor from '../../components/editor/TwinEditor'
 import EmptyFigure from '../../components/EmptyFigure'
 import Error from '../../components/Error'
-import IconsToolbar from '../../components/new-editor/IconsToolbar'
-import TwinEditor from '../../components/new-editor/TwinEditor'
 import Spinner from '../../components/Spinner'
 import useDebouncedCallback from '../../hooks/use-debounced-callback'
 import type { JsonValue } from '../../types/json'
