--- src/utils/autoformat.editor.ts
@@ -1,163 +1,0 @@
-import {
-  autoformatArrow,
-  autoformatLegal,
-  autoformatLegalHtml,
-  autoformatMath,
-  autoformatPunctuation,
-  autoformatSmartQuotes,
-  type AutoformatRule,
-} from '@udecode/plate-autoformat'
-import {
-  MARK_BOLD,
-  MARK_CODE,
-  MARK_ITALIC,
-  MARK_STRIKETHROUGH,
-  MARK_UNDERLINE,
-} from '@udecode/plate-basic-marks'
-import { ELEMENT_BLOCKQUOTE } from '@udecode/plate-block-quote'
-import { ELEMENT_CODE_BLOCK } from '@udecode/plate-code-block'
-import {
-  ELEMENT_H1,
-  ELEMENT_H2,
-  ELEMENT_H3,
-  ELEMENT_H4,
-  ELEMENT_H5,
-  ELEMENT_H6,
-} from '@udecode/plate-heading'
-import { ELEMENT_LI, ELEMENT_OL, ELEMENT_UL } from '@udecode/plate-list'
-import { formatCodeBlock, formatList, preFormat } from './format.editor'
-
-const autoformatBlocks: AutoformatRule[] = [
-  {
-    mode: 'block',
-    type: ELEMENT_H1,
-    match: '# ',
-    preFormat,
-  },
-  {
-    mode: 'block',
-    type: ELEMENT_H2,
-    match: '## ',
-    preFormat,
-  },
-  {
-    mode: 'block',
-    type: ELEMENT_H3,
-    match: '### ',
-    preFormat,
-  },
-  {
-    mode: 'block',
-    type: ELEMENT_H4,
-    match: '#### ',
-    preFormat,
-  },
-  {
-    mode: 'block',
-    type: ELEMENT_H5,
-    match: '##### ',
-    preFormat,
-  },
-  {
-    mode: 'block',
-    type: ELEMENT_H6,
-    match: '###### ',
-    preFormat,
-  },
-  {
-    mode: 'block',
-    type: ELEMENT_BLOCKQUOTE,
-    match: '> ',
-    preFormat,
-  },
-  {
-    mode: 'block',
-    type: ELEMENT_CODE_BLOCK,
-    match: '```',
-    triggerAtBlockStart: false,
-    preFormat,
-    format: (editor) => formatCodeBlock(editor),
-  },
-]
-
-export const autoformatMarks: AutoformatRule[] = [
-  {
-    mode: 'mark',
-    type: [MARK_BOLD, MARK_ITALIC],
-    match: '***',
-  },
-  {
-    mode: 'mark',
-    type: [MARK_UNDERLINE, MARK_ITALIC],
-    match: '__*',
-  },
-  {
-    mode: 'mark',
-    type: [MARK_UNDERLINE, MARK_BOLD],
-    match: '__**',
-  },
-  {
-    mode: 'mark',
-    type: [MARK_UNDERLINE, MARK_BOLD, MARK_ITALIC],
-    match: '___***',
-  },
-  {
-    mode: 'mark',
-    type: MARK_BOLD,
-    match: '**',
-  },
-  {
-    mode: 'mark',
-    type: MARK_UNDERLINE,
-    match: '__',
-  },
-  {
-    mode: 'mark',
-    type: MARK_ITALIC,
-    match: '*',
-  },
-  {
-    mode: 'mark',
-    type: MARK_ITALIC,
-    match: '_',
-  },
-  {
-    mode: 'mark',
-    type: MARK_STRIKETHROUGH,
-    match: '~~',
-  },
-  {
-    mode: 'mark',
-    type: MARK_CODE,
-    match: '`',
-  },
-]
-
-const autoformatLists: AutoformatRule[] = [
-  {
-    mode: 'block',
-    type: ELEMENT_LI,
-    match: ['* ', '- ', '+ '],
-    preFormat,
-    format: (editor) => formatList(editor, ELEMENT_UL),
-  },
-  {
-    mode: 'block',
-    type: ELEMENT_LI,
-    match: ['1. ', '1) '],
-    preFormat,
-    format: (editor) => formatList(editor, ELEMENT_OL),
-  },
-]
-
-export const autoformatRules = [
-  ...autoformatBlocks,
-  ...autoformatMarks,
-  ...autoformatLists,
-  ...autoformatSmartQuotes,
-  ...autoformatPunctuation,
-  ...autoformatLegal,
-  ...autoformatLegalHtml,
-  ...autoformatArrow,
-  ...autoformatMath,
-]
